requirements.txt
python-telegram-bot==20.7
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.2
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
